+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 
 ## vN.N.N
+* Add winking face emoji
 * Editing a slug during creation modal will result in a blank slug when hitting ENTER
 * Add More Fields On Designated Screens (display_title, credit_url, cta_text, cta_url, etc.)
 * Uploading a new pdf to a content block does not auto select or show up